requirements.txt
# Core Dependencies
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.3

# Machine Learning & Embeddings
sentence-transformers>=5.0.0
faiss-cpu==1.7.4
torch>=2.0.0  # Required by sentence-transformers
numpy==1.24.3

# Configuration
pyyaml>=6.0.0

# Utilities
python-multipart==0.0.6
requests==2.31.0
psutil>=5.9.0

# Testing
pytest==7.4.4
pytest-asyncio==0.21.1
httpx==0.25.2  # For testing FastAPI